+++ src/apis/base.ts
@@ -137,7 +137,7 @@ export function MustSubFS(fsys: string | FsRoot, dir: string): FsRoot {
 
 // Static is a handler function to serve static directory content from fsys.
 //
-// If a file resource is missing and indexFallback is set, the request
+// If a file resource is missing and indexFallback is true, the request
 // will be forwarded to the base index.html (useful for SPA with pretty urls).
 //
 // NB! Expects the route to have a "{path...}" wildcard parameter.
@@ -145,7 +145,7 @@ export function MustSubFS(fsys: string | FsRoot, dir: string): FsRoot {
 // Special redirects:
 //   - if "path" is a file that ends in index.html, it is redirected to its non-index.html version (eg. /test/index.html -> /test/)
 //   - if "path" is a directory that has index.html, the index.html file is rendered,
-//     otherwise if missing - returns 404 or fallback to the root index.html if indexFallback is set
+//     otherwise if missing - returns 404 or fallback to the root index.html if indexFallback is true
 export function Static(fsys: string | FsRoot, indexFallback: boolean): (event: RequestEvent) => Promise<unknown> {
   if (!fsys) {
     throw new Error("Static: the provided fs.FS argument is nil");